requirements.txt
requests~=2.31.0
pyspark~=3.5.0
apache-airflow-providers-apache-spark==4.6.0
python-dotenv~=1.0.0
boto3~=1.29.4